requirements.txt
requests
python-dotenv
kafka-python
confluent-kafka